Move test metadata test util
